busybox: scripts: add script that generates N distinct randconfigs
To compare the configuration (.config) files between each other, this script
does a checksum of each one of them (after their generation) with `cksum` and
saves it in a file (.db). Hence the checking is verifying if the checksum of the
lastly generated .config is in .db.

Since the date of the generation of the .config is written on Line 4 of each
file, the script removes Line 4 of each .config then does a checksum of the file
without this line.